// fuel/sentio-fuel/src/schema/schema.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type { String, Int, BigInt, Float, ID, Bytes, Timestamp, Boolean, Int8 } from '@sentio/sdk/store'
import { Entity, Required, One, Many, Column, ListColumn, AbstractEntity, getStore, UpdateValues } from '@sentio/sdk/store'
import { BigDecimal } from '@sentio/bigdecimal'
import { DatabaseSchema } from '@sentio/sdk'







interface L1BalanceConstructorInput {
  id: String;
  blockHeight: BigInt;
  timestamp: String;
  assetId: String;
  distinctId: String;
  amount: BigInt;
}
@Entity("L1Balance")
export class L1Balance extends AbstractEntity  {

	@Required
	@Column("String")
	id: String

	@Required
	@Column("BigInt")
	blockHeight: BigInt

	@Required
	@Column("String")
	timestamp: String

	@Required
	@Column("String")
	assetId: String

	@Required
	@Column("String")
	distinctId: String

	@Required
	@Column("BigInt")
	amount: BigInt
  constructor(data: L1BalanceConstructorInput) {super()}
  
  
  static update(values: UpdateValues<L1BalanceConstructorInput>): Promise<void> {
    return getStore().update(L1Balance, values)
  }
}


interface L2BalanceConstructorInput {
  id: String;
  blockHeight?: BigInt;
  daHeight?: BigInt;
  timestamp: String;
  assetId: String;
  distinctId: String;
  amount: BigInt;
}
@Entity("L2Balance")
export class L2Balance extends AbstractEntity  {

	@Required
	@Column("String")
	id: String

	@Column("BigInt")
	blockHeight?: BigInt

	@Column("BigInt")
	daHeight?: BigInt

	@Required
	@Column("String")
	timestamp: String

	@Required
	@Column("String")
	assetId: String

	@Required
	@Column("String")
	distinctId: String

	@Required
	@Column("BigInt")
	amount: BigInt
  constructor(data: L2BalanceConstructorInput) {super()}
  
  
  static update(values: UpdateValues<L2BalanceConstructorInput>): Promise<void> {
    return getStore().update(L2Balance, values)
  }
}


const source = `type L1Balance @entity {
  id: String!
  blockHeight: BigInt!
  timestamp: String!
  assetId: String!
  distinctId: String!
  amount: BigInt!
}

type L2Balance @entity {
  id: String!
  blockHeight: BigInt
  daHeight: BigInt
  timestamp: String!
  assetId: String!
  distinctId: String!
  amount: BigInt!
}
`
DatabaseSchema.register({
  source,
  entities: {
    "L1Balance": L1Balance,
		"L2Balance": L2Balance
  }
})